+++ Toggly.FeatureManagement.Python/toggly/tests/telemetry/__init__.py
@@ -1,0 +1,1 @@
+"""Telemetry test package."""